Add back modifyApproval to Mutations and rename support function file
